feat: refactor tray setup and integrate resource paths for icons
