+++ src/core/bootstrap/register-app-ready-flow.ts
@@ -3,7 +3,7 @@ import { logger } from '@internxt/drive-desktop-core/build/backend';
 import eventBus from '../../apps/main/event-bus';
 import { getIsLoggedIn } from '../../apps/main/auth/handlers';
 import { createAuthWindow } from '../../apps/main/windows/auth';
-import { setTrayStatus } from '../../apps/main/tray/tray';
+import { setupTrayIcon, setTrayStatus } from '../../apps/main/tray/tray-setup';
 import { broadcastToWindows } from '../../apps/main/windows';
 import { setupThemeListener } from '../theme';
 import { registerAvailableUserProductsHandlers } from '../../backend/features/payments/ipc/register-available-user-products-handlers';
@@ -33,6 +33,7 @@ export function registerAppReadyFlow() {
        */
       // await installNautilusExtension();
       setupThemeListener();
+      setupTrayIcon();
 
       eventBus.emit('APP_IS_READY');
       const isLoggedIn = getIsLoggedIn();
