+++ src/core/electron/paths.ts
@@ -18,6 +18,9 @@ const FUSE_DAEMON_SOCKET = join(process.env.XDG_RUNTIME_DIR ?? '/tmp', 'internxt
 const FUSE_DAEMON_BINARY = app.isPackaged
   ? join(process.resourcesPath, 'fuse-daemon')
   : join(__dirname, '../../../dist/fuse-daemon');
+const RESOURCES_PATH = app.isPackaged
+  ? path.join(process.resourcesPath, 'assets')
+  : path.join(__dirname, '../../../assets');
 
 export const PATHS = {
   HOME_FOLDER_PATH,
@@ -32,4 +35,5 @@ export const PATHS = {
   FUSE_DAEMON_LOG,
   FUSE_DAEMON_SOCKET,
   FUSE_DAEMON_BINARY,
+  RESOURCES_PATH,
 };
